Updating emails

+++ src/transform/transform.ts
@@ -26,7 +26,7 @@ export const BuildGenericContent= (content: string, club: core.Club.Model): Rich
  */
 export const BuildEventContent = (event: core.Event.Model, club: core.Club.Model, link: string): EventInfo => {
     const date = new Date(event.start)
-    var dateOptions = { weekday: 'long', year: 'numeric', month: 'long', day: 'numeric' };
+    var dateOptions = { weekday: 'short', year: 'numeric', month: 'long', day: 'numeric' };
     var timeOptions = { hour: 'numeric', minute: 'numeric' }
     const richContent: EventInfo = {
         name: event.name,
